+++ src/mdapy/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2022-2025, Yongchao Wu in Aalto University
+# Copyright (c) 2022-2026, Yongchao Wu in Aalto University
 # This file is from the mdapy project, released under the BSD 3-Clause License.
 
 from mdapy.system import System
@@ -42,4 +42,4 @@ __all__ = [
     "LindemannParameter",
     "VoidAnalysis",
 ]
-__version__ = "1.0.4a1"
+__version__ = "1.0.4a2"