+++ src/warren_cowley_parameter.cpp
@@ -1,3 +1,5 @@
+// Copyright (c) 2022-2026, Yongchao Wu in Aalto University
+// This file is from the mdapy project, released under the BSD 3-Clause License.
 #include "type.h"
 #include <nanobind/nanobind.h>
 #include <nanobind/ndarray.h>