+++ tests/test_PCA.py
@@ -1,3 +1,5 @@
+# Copyright (c) 2022-2026, Yongchao Wu in Aalto University
+# This file is from the mdapy project, released under the BSD 3-Clause License.
 from sklearn.decomposition import PCA as PCA_S
 from mdapy.potential_tool import PCA
 import numpy as np