+++ tests/test_angular_distribution_function.py
@@ -1,3 +1,5 @@
+# Copyright (c) 2022-2026, Yongchao Wu in Aalto University
+# This file is from the mdapy project, released under the BSD 3-Clause License.
 from mdapy import System
 from ovito.modifiers import BondAnalysisModifier, CreateBondsModifier
 import numpy as np