+++ tests/test_centro_symmetry_parameter.py
@@ -1,3 +1,5 @@
+# Copyright (c) 2022-2026, Yongchao Wu in Aalto University
+# This file is from the mdapy project, released under the BSD 3-Clause License.
 import mdapy as mp
 from ovito.modifiers import CentroSymmetryModifier
 import numpy as np