+++ tests/test_msd.py
@@ -1,3 +1,5 @@
+# Copyright (c) 2022-2026, Yongchao Wu in Aalto University
+# This file is from the mdapy project, released under the BSD 3-Clause License.
 import numpy as np
 from mdapy.mean_squared_displacement import MeanSquaredDisplacement
 import freud